refactor: global 위치의 baseEntity import
