+++ src/main/java/com/project/likelion14thbe/domain/order/entity/Order.java
@@ -1,6 +1,7 @@
 package com.project.likelion14thbe.domain.order.entity;
 
 import com.project.likelion14thbe.domain.member.entity.Member;
+import com.project.likelion14thbe.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -12,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Table(name = "orders")
-public class Order extends BaseEntity{
+public class Order extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
